+++ modules/contrib/blazy/js/src/base/io/bio.media.js
@@ -165,7 +165,6 @@
       }
     }
 
-
     me.erCount = ER_COUNTED;
   };